doc_app/requirements.txt

**requirements.txt:**
```plaintext
Pillow==8.3.2
pandas==1.3.3
openpyxl==3.1.2
